ejr-3/requirements.txt
Django==5.0.6
python-dotenv==1.0.1
google-generativeai==0.8.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas
numpy
scikit-learn
plotly
dash
dash-bootstrap-components
requests
matplotlib
seaborn




